refactor: 관측 관련 설정 파일을 별도 모듈로 분리
