+++ common/build.gradle
@@ -19,11 +19,4 @@ dependencies {
 
     // [OpenApi]
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-api'
-
-    // [Observability - Logging]
-    api 'net.logstash.logback:logstash-logback-encoder:8.0'
-
-    // [Observability - Metrics]
-    api 'org.springframework.boot:spring-boot-starter-actuator'
-    api 'io.micrometer:micrometer-registry-prometheus'
 }
