--- common/src/main/java/com/truve/platform/common/config/MetricsConfig.java
@@ -1,19 +1,0 @@
-package com.truve.platform.common.config;
-
-import io.micrometer.core.instrument.MeterRegistry;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.actuate.autoconfigure.metrics.MeterRegistryCustomizer;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class MetricsConfig {
-
-    @Value("${spring.application.name:unknown-service}")
-    private String applicationName;
-
-    @Bean
-    MeterRegistryCustomizer<MeterRegistry> metricsCommonTags() {
-        return registry -> registry.config().commonTags("application", applicationName);
-    }
-}
