+++ settings.gradle
@@ -1,5 +1,6 @@
 rootProject.name = 'truve-backend'
 include ':common'
+include ':common-observability'
 include ':auth-server'
 include ':api-gateway'
 include ':payment'
